(extension) restore popup route via PUSH so back button works [DA-534]
The index loader now unconditionally redirects to /mount; persisted-route
restoration moves to a one-shot effect in RootRouter that calls
router.navigate after mount. react-router's loader redirects REPLACE
history, which left the in-page back button dead-ending on the restored
page. Restoring via a post-mount navigate produces a PUSH on top of
/mount, so back() returns to /mount as expected.
